requirements.txt
Django==4.2.7
graphene-django==3.1.5
pymongo==4.6.1
dnspython==2.4.2
nltk==3.8.1
scikit-learn==1.5.2
numpy==2.1.3
pandas==2.2.3
joblib==1.4.2
django-cors-headers==4.3.1
python-dotenv==1.0.0
whitenoise==6.6.0
certifi==2024.2.2
